pyproject.toml
[project]
name = "kushim"
version = "0.2.0"
description = "A framework for building high-quality, verifiable evaluation datasets for LLMs."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "polars>=0.20.0",
    "dspy>=2.4.5",
    "wikipedia>=1.4.0",
    "python-dotenv>=1.0.0",
    "llama-index",
    "datasets>=3.6.0",
    "huggingface-hub[cli]>=0.33.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# This table tells the build backend (hatchling) that the `kushim`
# directory is the package to be included in the build distribution.
[tool.hatch.build.targets.wheel]
packages = ["kushim"]

[tool.pytest.ini_options]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-aiohttp>=1.0.0",
    "polars[testing]>=0.20.0",
]
